+++ client/package.json
@@ -20,7 +20,7 @@
     "graphql.macro": "^1.4.2",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
-    "react-scripts": "3.3.1",
+    "react-scripts": "^3.4.1",
     "styled-components": "^5.0.1",
     "subscriptions-transport-ws": "^0.9.18"
   },
